move linked apps task to tasks.py
